% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{check_aov_inputs}
\alias{check_aov_inputs}
\title{Validate inputs for optim_aov (in Shiny)}
\usage{
check_aov_inputs(
  N,
  levels,
  target_group_means,
  range,
  tolerance,
  max_iter,
  init_temp,
  cooling_rate,
  max_starts,
  max_step,
  parallel_start
)
}
\value{
TRUE if all inputs are valid; otherwise shows an error notification
  and returns FALSE.
}
\description{
Validate inputs for optim_aov (in Shiny)
}
