% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_grim.R
\name{check_grim}
\alias{check_grim}
\title{Perform GRIM test on a mean}
\usage{
check_grim(n, target_mean, decimals, tol.r = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Integer sample size}

\item{target_mean}{Numeric target mean to test}

\item{decimals}{Integer number of decimal places}

\item{tol.r}{Numeric tolerance for rounding errors}
}
\value{
A list with components:
  - test: Logical, TRUE if plausible
  - grim_mean: Numeric, plausible mean
}
\description{
Perform GRIM test on a mean
}
