% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{check_lm_inputs}
\alias{check_lm_inputs}
\title{Validate inputs for optim_lm (in Shiny)}
\usage{
check_lm_inputs(
  tolerance,
  max_iter,
  init_temp,
  cooling_rate,
  hill_climbs,
  max_starts,
  parallel_start
)
}
\value{
TRUE if all inputs are valid; otherwise shows an error notification
  and returns FALSE.
}
\description{
Validate inputs for optim_lm (in Shiny)
}
