% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discourse_users_list.R
\name{discourse_users_list}
\alias{discourse_users_list}
\title{discourse_users_list}
\usage{
discourse_users_list(base_url)
}
\arguments{
\item{base_url}{The base URL of your Discourse instance}
}
\description{
Get a list of all Discourse users
}
