% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_rmse.R
\name{get_rmse}
\alias{get_rmse}
\title{Compute RMSE for a single discourse.object result}
\usage{
get_rmse(result)
}
\arguments{
\item{result}{A discourse.object with optimized data and inputs}
}
\value{
Named list of RMSE values
}
\description{
Compute RMSE for a single discourse.object result
}
