% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_rmse_parallel.R
\name{get_rmse_parallel}
\alias{get_rmse_parallel}
\title{Compute RMSE metrics across discourse.object runs}
\usage{
get_rmse_parallel(object_list)
}
\arguments{
\item{object_list}{A discourse.object or list thereof}
}
\value{
List with between-run and target RMSE summaries and raw values
}
\description{
Compute RMSE metrics across discourse.object runs
}
