% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Extract statistics from a single discourse.object}
\usage{
get_stats(result)
}
\arguments{
\item{result}{A discourse.object with data and inputs}
}
\value{
List containing model fit and summary stats or ANOVA/F stats or vectors
}
\description{
Extract statistics from a single discourse.object
}
