% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_stats_parallel.R
\name{get_stats_parallel}
\alias{get_stats_parallel}
\title{Aggregate stats across discourse.object runs}
\usage{
get_stats_parallel(object_list)
}
\arguments{
\item{object_list}{List of discourse.object}
}
\value{
List of data.frames with mean, median, sd, min, max per component
}
\description{
Aggregate stats across discourse.object runs
}
