% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optim_lme.R
\name{optim_lme}
\alias{optim_lme}
\title{optim_lme()}
\usage{
optim_lme(
  sim_data,
  target_cor,
  target_reg,
  reg_equation,
  target_se = NULL,
  weight = c(1, 1),
  max_iter = 1e+05,
  init_temp = 1,
  cooling_rate = NULL,
  tol = 1e-06,
  progress_bar = TRUE,
  max_starts = 1,
  hill_climbs = NULL,
  move_prob = list(start = c(residual = 0, k_cycle = 0, local = 0.25, tau = 0.75), end =
    c(residual = 0.2, k_cycle = 0.1, local = 0.7, tau = 0)),
  min_decimals = 0,
  eps = 1e-05
)
}
\arguments{
\item{sim_data}{Data frame (wide) of predictors and outcome.}

\item{target_cor}{Numeric vector of target correlations (upper‐tri of full corr matrix).}

\item{target_reg}{Numeric vector of target regression coefficients.}

\item{reg_equation}{Model formula string, e.g. "Y ~ X1 + X2 + X1:X2".}

\item{target_se}{Optional target standard errors for regression.}

\item{weight}{Numeric length‐2 vector of weights for (corr, reg) error.}

\item{max_iter}{Maximum SA iterations per start.}

\item{init_temp}{Initial temperature.}

\item{cooling_rate}{Cooling rate per iteration (if NA, set to (max_iter-10)/max_iter).}

\item{tol}{Error tolerance for early stopping.}

\item{progress_bar}{Logical; show progress bar.}

\item{max_starts}{Number of SA restarts.}

\item{hill_climbs}{Optional hill‐climb iterations after SA.}

\item{move_prob}{List containing start and end probabilities of move selection}

\item{min_decimals}{Integer; then minimum number of decimals (trailing zeros) of the targets}

\item{eps}{Numeric; the lower bound for the standardization to prevent division by zero}
}
\value{
A "discourse.object" with best error, data, inputs, and trace
}
\description{
Optimize a longitudinal mixed‐model data set via simulated annealing.
}
