% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parallel_lm.R
\name{parallel_lm}
\alias{parallel_lm}
\title{Parallel Linear Model Optimization}
\usage{
parallel_lm(
  parallel_start,
  return_best_solution,
  sim_data,
  target_cor,
  target_reg,
  reg_equation,
  target_se = NULL,
  weight = c(1, 1),
  max_iter = 1e+05,
  init_temp = 1,
  cooling_rate = NULL,
  tol = 1e-06,
  prob_global_move = 0.1,
  max_starts = 1,
  hill_climbs = NULL,
  min_decimals = 1
)
}
\arguments{
\item{parallel_start}{Number of parallel runs.}

\item{return_best_solution}{Logical; return best solution only.}

\item{sim_data}{Simulation data.}

\item{target_cor}{Target correlation.}

\item{target_reg}{Target regression.}

\item{reg_equation}{Regression equation.}

\item{target_se}{Optional standard error target.}

\item{weight}{Numeric vector of length 2: run weights.}

\item{max_iter}{Maximum number of iterations.}

\item{init_temp}{Initial temperature.}

\item{cooling_rate}{Cooling schedule rate.}

\item{tol}{Convergence tolerance.}

\item{prob_global_move}{Probability of global move.}

\item{max_starts}{Number of starts per run.}

\item{hill_climbs}{Number of hill climbs.}

\item{min_decimals}{Integer; then minimum number of decimals (trailing zeros) of the targets}
}
\value{
List of optimization results or single best result.
}
\description{
Run parallelized optimization for linear model parameters.
}
