% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_cooling.R
\name{plot_cooling}
\alias{plot_cooling}
\title{Plot cooling schedule of a \code{discourse.object}}
\usage{
plot_cooling(discourse_obj)
}
\arguments{
\item{discourse_obj}{Output from \code{optim_*} functions, with class \code{discourse.object}.}
}
\value{
A \pkg{ggplot2} line plot of squared temperature vs. iteration number.
}
\description{
Plot cooling schedule of a \code{discourse.object}
}
