% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rmse.R
\name{plot_rmse}
\alias{plot_rmse}
\title{Plot RMSE Comparison}
\usage{
plot_rmse(object_list)
}
\arguments{
\item{object_list}{List of discourse objects.}
}
\value{
ggplot2 object
}
\description{
Plot RMSE distributions (Between vs. Target) across metrics.
}
