% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{summary.discourse.object}
\alias{summary.discourse.object}
\title{Summarize a discourse.object}
\usage{
\method{summary}{discourse.object}(object, ...)
}
\arguments{
\item{object}{A `discourse.object` to summarize}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `summary.discourse.object`
}
\description{
Summarize a discourse.object
}
