% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_weights_est.R
\name{weights_est}
\alias{weights_est}
\title{Estimate simulation weights via simulated annealing}
\usage{
weights_est(
  module,
  sim_runs,
  sim_data,
  target_cor,
  target_reg,
  target_se = NULL,
  reg_equation,
  max_iter = 1e+05,
  init_temp = 1,
  cooling_rate = NULL,
  tol = 1e-06,
  prob_global_move = 0.05,
  progress_bar = TRUE,
  weight = c(1, 1),
  pool_range = 10,
  max_starts = 1,
  parallel_start = 1,
  move_prob = list(start = c(residual = 0, k_cycle = 0, local = 0.25, tau = 0.75), end =
    c(residual = 0.2, k_cycle = 0.1, local = 0.7, tau = 0)),
  min_decimals = 0
)
}
\arguments{
\item{module}{Character; either `"lm"` or `"lme"` to select the model type.}

\item{sim_runs}{Integer; number of simulation runs.}

\item{sim_data}{Data frame or matrix of predictors and outcome.}

\item{target_cor}{Numeric vector of target correlations.}

\item{target_reg}{Numeric vector of target regression coefficients.}

\item{target_se}{Numeric vector of target standard errors (optional).}

\item{reg_equation}{Character string of the regression formula.}

\item{max_iter}{Integer; maximum number of iterations (default `1e5`).}

\item{init_temp}{Numeric; initial temperature for simulated annealing.}

\item{cooling_rate}{Numeric; per‑iteration cooling rate.}

\item{tol}{Numeric; convergence tolerance.}

\item{prob_global_move}{Numeric; probability of taking a global move.}

\item{progress_bar}{Logical; if `TRUE`, print progress.}

\item{weight}{Numeric vector; initial weights (default `c(1,1)`).}

\item{pool_range}{Integer; how many recent error ratios to average.}

\item{max_starts}{Integer; number of restarts.}

\item{parallel_start}{Integer; number of parallel workers (0 = sequential).}

\item{move_prob}{List containing start and end probabilities of move selection}

\item{min_decimals}{Integer; then minimum number of decimals (trailing zeros) of the targets}
}
\value{
A list with components:
\describe{
  \item{weights}{Numeric vector of estimated weights.}
  \item{best_solution}{Best solution object from the last optimizer run.}
}
}
\description{
Run multiple simulated‑annealing optimizations (LM or LME) to find weights that
reproduce target correlations and regression coefficients.
}
