% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discourse_auth.R
\name{discourse_auth}
\alias{discourse_auth}
\title{discourse_auth}
\usage{
discourse_auth(account_id, api_name)
}
\arguments{
\item{api_name}{The Discourse account name}

\item{api_key}{The API key}
}
\description{
Authenticate your Discourse API account
}
