% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hill_climb.R
\name{hill_climb}
\alias{hill_climb}
\title{Perform hill-climbing optimization}
\usage{
hill_climb(
  current_candidate,
  error_function,
  N,
  hill_climbs = 100,
  LME = FALSE,
  num_preds = NULL,
  progress_bar = TRUE,
  neighborhood_size = 4,
  w.length = NULL,
  outcome = NULL
)
}
\arguments{
\item{current_candidate}{Initial candidate data structure}

\item{error_function}{Function returning list with element $total_error}

\item{N}{The sample size}

\item{hill_climbs}{Maximum iterations (default 1e4)}

\item{LME}{Logical, use mixed-model moves if TRUE}

\item{num_preds}{Number of predictors for standard models}

\item{progress_bar}{Show progress bar if TRUE}

\item{neighborhood_size}{Moves per iteration}

\item{w.length}{The length of the LME data in wide format}

\item{outcome}{The outcome vector in the LM module}
}
\value{
List with best_candidate and best_error
}
\description{
Perform hill-climbing optimization
}
