% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optim_vec.R
\name{optim_vec}
\alias{optim_vec}
\title{Optimize a vector matrix to match target means and SDs}
\usage{
optim_vec(
  N,
  target_mean,
  target_sd,
  range,
  tolerance,
  integer,
  max_iter = 1e+05,
  init_temp = 1,
  cooling_rate = NA,
  int.probs = NULL,
  pb_update_interval = NA,
  obj_weight = list(c(1, 1)),
  maxit_pso = 1000,
  eps = 1e-12,
  max_starts = 3,
  checkGrim = TRUE,
  prob_heuristic = 0.1,
  parallel = FALSE
)
}
\arguments{
\item{N}{Integer, number of values.}

\item{target_mean}{Named numeric vector of desired means.}

\item{target_sd}{Named numeric vector of desired SDs.}

\item{range}{Numeric of length 2 or matrix; allowed value ranges.}

\item{tolerance}{Numeric; stop when objective ≤ tolerance.}

\item{integer}{Logical; use integer values if TRUE.}

\item{max_iter}{Integer; max annealing iterations (default 1e5).}

\item{init_temp}{Numeric; starting temperature.}

\item{cooling_rate}{Numeric; decay rate (default auto).}

\item{int.probs}{Numeric; a list with a vector of sampling probabilities for each variable.}

\item{pb_update_interval}{Integer; progress‐bar refresh frequency.}

\item{obj_weight}{List of two numerics; weights for mean vs SD.}

\item{maxit_pso}{Integer; max PSO iterations (continuous mode).}

\item{eps}{Numeric; small constant to avoid zero division.}

\item{max_starts}{Integer; annealing restarts.}

\item{checkGrim}{Logical; run GRIM check for integers.}

\item{prob_heuristic}{Numeric; heuristic move probability.}

\item{parallel}{Logical; run per-variable in parallel.}
}
\value{
A `discourse.object` with:
* `data`: matrix of optimized vectors.
* `best_error`: list of final errors.
* `track_error`: list of error paths.
* `inputs`: list of input parameters.
}
\description{
Optimize a vector matrix to match target means and SDs
}
